Organize tooling and EEG ICA wrappers
